FishGroup add Scene change but FishGroupManager is a bad name it should call SceneManager

+++ Src/FM79979Engine/Core/FatmingAI/FAIStatus.cpp
@@ -73,7 +73,7 @@ namespace FATMING_AI
 			if(l_pCharacterAttributeAndUIInfo->CharacterAttribute == e_CharacterAttribute)
 			{
 				if( e_piReceivedIndex )
-					*e_piReceivedIndex = i;
+					*e_piReceivedIndex = (int)i;
 				return l_pCharacterAttributeAndUIInfo;
 			}
 		}
